+++ dwara-commons/src/main/java/org/ishafoundation/dwaraapi/ProcessingTaskNames.java
@@ -1,0 +1,8 @@
+package org.ishafoundation.dwaraapi;
+
+public interface ProcessingTaskNames {
+
+	public static final String PROCESSING_TASK_4K_TO_HD = "video-4k-hd-gen";
+}
+
+